architecture refactoring
