+++ src/main/java/com/tribesbackend/tribes/tribesresources/service/ResourcesModelHelpersMethods.java
@@ -1,4 +1,5 @@
-package com.tribesbackend.tribes.tribesresources.model;
+package com.tribesbackend.tribes.tribesresources.service;
+import com.tribesbackend.tribes.tribesresources.model.ResourcesModel;
 import com.tribesbackend.tribes.tribesresources.repository.ResourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
